apps/worker/requirements.txt
redis>=5.0.0
unstructured-client>=0.40.0
openai>=1.0.0
pgvector>=0.2.0
psycopg2-binary>=2.9.0
